# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Upcreate Creator Intelligence API",
    "description": "Backend API for creator metrics, commerce scoring, brands, and campaigns (MVP).",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "/api/v1", "description": "Relative (same host)" },
    { "url": "http://localhost:4000/api/v1", "description": "Local dev" }
  ],
...
  "paths": {
...
    "/campaigns": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "List campaigns",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          {
            "name": "status",
            "in": "query",
            "schema": { "$ref": "#/components/schemas/CampaignStatus" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CampaignListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Campaigns"],
        "summary": "Create campaign",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCampaign" }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Campaign" }
              }
            }
          },
          "404": { "description": "Brand not found" }
        }
      }
    },
    "/campaigns/{id}": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "Get campaign",
        "parameters": [{ "$ref": "#/components/parameters/CampaignId" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CampaignDetail" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "tags": ["Campaigns"],
        "summary": "Update campaign",
        "parameters": [{ "$ref": "#/components/parameters/CampaignId" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCampaign" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Campaign" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      }
    },
    "/campaigns/{id}/status": {
      "patch": {
        "tags": ["Campaigns"],
        "summary": "Transition campaign status",
        "description": "Top-level campaign only: DRAFT→ACTIVE; ACTIVE↔PAUSED; ACTIVE|PAUSED→COMPLETED|CANCELLED. Pairing WMS is on campaign-creators.",
        "parameters": [{ "$ref": "#/components/parameters/CampaignId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "$ref": "#/components/schemas/CampaignStatus" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Campaign" }
              }
            }
          },
          "400": { "description": "Invalid transition" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/campaigns/{id}/invite": {
      "post": {
        "tags": ["Campaigns"],
        "summary": "Invite / recommend creator (ADMIN JWT)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/CampaignId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["creatorId"],
                "properties": { "creatorId": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Pairing created (AWAITING_CREATOR)" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Not admin" },
          "404": { "description": "Campaign or creator not found" },
          "409": { "description": "Already paired" }
        }
      }
    },
    "/campaigns/{id}/recommend": {
      "post": {
        "tags": ["Campaigns"],
        "summary": "Admin recommends creator (same as invite; ACTIVE campaign only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/CampaignId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["creatorId"],
                "properties": { "creatorId": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Pairing created" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Not admin" }
        }
      }
    },
    "/campaigns/{id}/affiliate-summary": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "Affiliate clicks/conversions summary for campaign",
        "parameters": [{ "$ref": "#/components/parameters/CampaignId" }],
        "responses": { "200": { "description": "Summary object" } }
      }
    },
...
    "/campaign-creators/{id}/creator-response": {
      "patch": {
        "tags": ["CampaignCreators"],
        "summary": "Creator accept/decline (AWAITING_CREATOR)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["accept"],
                "properties": { "accept": { "type": "boolean" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated pairing" } }
      }
    },
    "/campaign-creators/{id}/brand-response": {
      "patch": {
        "tags": ["CampaignCreators"],
        "summary": "Brand accept/decline (AWAITING_BRAND)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["accept"],
                "properties": { "accept": { "type": "boolean" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated pairing" } }
      }
    },
    "/campaign-creators/{id}/contents": {
      "get": {
        "tags": ["CampaignCreators"],
        "summary": "List content submissions for pairing",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "List" } }
      },
      "post": {
        "tags": ["CampaignCreators"],
        "summary": "Submit content metadata (creator JWT)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "storageUrl": { "type": "string", "format": "uri" },
                  "assetKey": { "type": "string" },
                  "mimeType": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Login via POST /auth/login; send Authorization: Bearer <token>"
      }
    },
    "parameters": {
...
      "CampaignId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      }
    },
    "schemas": {
...
      "CampaignStatus": {
        "type": "string",
        "enum": ["DRAFT", "ACTIVE", "PAUSED", "COMPLETED", "CANCELLED"]
      },
...
      "Campaign": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "brandId": { "type": "string" },
          "title": { "type": "string" },
          "status": { "$ref": "#/components/schemas/CampaignStatus" },
          "budget": { "type": "number", "nullable": true },
          "brief": { "type": "string", "nullable": true },
          "productInfo": { "type": "string", "nullable": true, "description": "JSON string or free text (MVP)" },
          "commissionRate": { "type": "number", "nullable": true, "description": "Percentage e.g. 12.5" },
          "startsAt": { "type": "string", "format": "date-time", "nullable": true },
          "endsAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CampaignDetail": {
        "type": "object",
        "description": "Campaign with brand and invited creators"
      },
      "CampaignListResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "type": "object" } },
          "meta": { "$ref": "#/components/schemas/PaginationMeta" }
        }
      },
      "CreateCampaign": {
        "type": "object",
        "required": ["brandId", "title"],
        "properties": {
          "brandId": { "type": "string" },
          "title": { "type": "string", "maxLength": 300 },
          "budget": { "type": "number", "exclusiveMinimum": 0 },
          "brief": { "type": "string" },
          "productInfo": { "type": "string" },
          "commissionRate": { "type": "number", "minimum": 0, "maximum": 100 },
          "startsAt": { "type": "string", "format": "date-time" },
          "endsAt": { "type": "string", "format": "date-time" }
        }
      },
      "UpdateCampaign": {
        "type": "object",
        "properties": {
          "brandId": { "type": "string" },
          "title": { "type": "string" },
          "budget": { "type": "number" },
          "brief": { "type": "string" },
          "productInfo": { "type": "string" },
          "commissionRate": { "type": "number", "minimum": 0, "maximum": 100 },
          "startsAt": { "type": "string", "format": "date-time" },
          "endsAt": { "type": "string", "format": "date-time" }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
    }
  }
}